Combine print and non print page
