+++ app/javascript/packs/component-definitions-common.js
@@ -173,7 +173,6 @@ import VmEditForm from '../components/vm-edit-form';
 import VmFloatingIPsForm from '../components/vm-floating-ips/vm-floating-ips-form';
 import VmResizeForm from '../components/vm-resize-form/vm-resize-form';
 import VmRightSize from '../components/vm-right-size';
-import VmRightSizePrint from '../components/vm-right-size/VmRightSizePrint';
 import VmServerRelationshipForm from '../components/vm-server-relationship-form';
 import VmSnapshotForm from '../components/vm-snapshot-form/vm-snapshot-form';
 import VmSnapshotTreeSelect from '../components/vm-snapshot-tree-select';
@@ -381,7 +380,6 @@ ManageIQ.component.addReact('VmEditForm', VmEditForm);
 ManageIQ.component.addReact('VmFloatingIPsForm', VmFloatingIPsForm);
 ManageIQ.component.addReact('VmResizeForm', VmResizeForm);
 ManageIQ.component.addReact('VmRightSize', VmRightSize);
-ManageIQ.component.addReact('VmRightSizePrint', VmRightSizePrint);
 ManageIQ.component.addReact('VmServerRelationshipForm', VmServerRelationshipForm);
 ManageIQ.component.addReact('VmSnapshotForm', VmSnapshotForm);
 ManageIQ.component.addReact('VmSnapshotTreeSelect', VmSnapshotTreeSelect);
